+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * 02f951e - 2017-08-27: Added code to flush episode objects from show objects created during Trakt sync 
+- * 755c973 - 2017-08-27: Switched Trakt sync code back to using database calls, faster 
+- * 2935e89 - 2017-08-27: Added code to flush episode objects from show objects created during Trakt sync 
 - * b6172d5 - 2017-08-27: Fixed startup issue do to unicode decode on os.walk 
 - * 0beea2b - 2017-08-27: Added arrow lib to requirements for Trakt API module 
 - * 27326ca - 2017-08-27: Moved code to core for removing stale .pyc files